added custom data source and made some fixes
- added "class" VXCustomDataSource allowing a user to write a custom data source
- increased warning level on MSVC to /W4 and resolved warnings arising from that
- added VX_UNUSED macro that allows shutting up unused-arg warnings explicitly when needed

+++ Bindings/Cpp/VXInstructionDecoder.hpp
@@ -323,7 +323,7 @@ inline uint8_t VXStreamDataSource::internalInputPeek()
     {
         return 0;
     }
-    return m_inputStream->peek();
+    return static_cast<uint8_t>(m_inputStream->peek());
 }
 
 inline uint8_t VXStreamDataSource::internalInputNext()
@@ -332,7 +332,7 @@ inline uint8_t VXStreamDataSource::internalInputNext()
     {
         return 0;
     }
-    return m_inputStream->get();
+    return static_cast<uint8_t>(m_inputStream->get());
 }
 
 inline bool VXStreamDataSource::isEndOfInput() const
